+++ frontend/src/views/SuplierView.vue
@@ -1,47 +1,11 @@
 <template>
   <!------------ALERT----------------->
-  <section>
-    <div v-if="errorMessage"
-      class="flex items-center w-full max-w-sm overflow-hidden bg-white rounded-lg shadow-md dark:bg-gray-800 absolute right-0">
-      <div class="flex items-center justify-center w-12 bg-red-500">
-        <svg class="w-12 h-12 text-white fill-current" viewBox="0 0 40 40" xmlns="http://www.w3.org/2000/svg">
-          <path
-            d="M20 3.36667C10.8167 3.36667 3.3667 10.8167 3.3667 20C3.3667 29.1833 10.8167 36.6333 20 36.6333C29.1834 36.6333 36.6334 29.1833 36.6334 20C36.6334 10.8167 29.1834 3.36667 20 3.36667ZM19.1334 33.3333V22.9H13.3334L21.6667 6.66667V17.1H27.25L19.1334 33.3333Z" />
-        </svg>
-      </div>
-
-      <div class="px-4 py-2 -mx-3">
-        <div class="mx-3">
-          <span class="font-semibold text-red-500 dark:text-red-400">Error</span>
-          <p class="text-sm text-gray-600 dark:text-gray-200">{{ errorMessage }}</p>
-        </div>
-      </div>
-    </div>
-
-    <div v-if="successMessage"
-      class="flex items-center w-full max-w-sm overflow-hidden bg-white rounded-lg shadow-md dark:bg-gray-800 absolute right-0">
-      <div class="flex items-center justify-center w-12 bg-emerald-500">
-        <svg class="w-6 h-6 text-white fill-current" viewBox="0 0 40 40" xmlns="http://www.w3.org/2000/svg">
-          <path
-            d="M20 3.33331C10.8 3.33331 3.33337 10.8 3.33337 20C3.33337 29.2 10.8 36.6666 20 36.6666C29.2 36.6666 36.6667 29.2 36.6667 20C36.6667 10.8 29.2 3.33331 20 3.33331ZM16.6667 28.3333L8.33337 20L10.6834 17.65L16.6667 23.6166L29.3167 10.9666L31.6667 13.3333L16.6667 28.3333Z" />
-        </svg>
-      </div>
-
-      <div class="px-4 py-2 -mx-3">
-        <div class="mx-3">
-          <span class="font-semibold text-red-500 dark:text-emerald-400">Success</span>
-          <p class="text-sm text-gray-600 dark:text-gray-200">{{ successMessage }}</p>
-        </div>
-      </div>
-    </div>
-
-  </section>
+  <TheAlert v-if="isAlertVisible" :message="alertStore.message" :type="alertStore.type" />
 
   <!------------FORM----------------->
   <template>
     <FormSuplier :suplierData="suplier" />
   </template>
-
 
   <!------------TABELA----------------->
   <section class="container px-4 py-6 mx-auto">
@@ -64,8 +28,11 @@
           </svg>
         </span>
 
-        <input type="text" placeholder="Search"
-          class="block w-full py-1.5 pr-5 text-gray-700 bg-white border border-gray-200 rounded-lg md:w-80 placeholder-gray-400/70 pl-11 rtl:pr-11 rtl:pl-5 dark:bg-gray-900 dark:text-gray-300 dark:border-gray-600 focus:border-blue-400 dark:focus:border-blue-300 focus:ring-blue-300 focus:outline-none focus:ring focus:ring-opacity-40">
+        <form>
+          <input type="text" placeholder="Pesquisar" @input="filterSuplier"
+            class="block w-full py-1.5 pr-5 text-gray-700 bg-white border border-gray-200 rounded-lg md:w-80 placeholder-gray-400/70 pl-11 rtl:pr-11 rtl:pl-5 dark:bg-gray-900 dark:text-gray-300 dark:border-gray-600 focus:border-blue-400 dark:focus:border-blue-300 focus:ring-blue-300 focus:outline-none focus:ring focus:ring-opacity-40">
+        </form>
+
       </div>
     </div>
 
@@ -79,7 +46,7 @@
                 <tr>
                   <th scope="col"
                     class="py-3.5 px-4 text-sm font-normal text-left rtl:text-right text-gray-500 dark:text-gray-400">
-                    <button class="flex items-center gap-x-3 focus:outline-none">
+                    <button class="flex items-center gap-x-3 focus:outline-none" @click="OrderingSuplier('name')">
                       <span>Nome Fantasia</span>
 
                       <svg class="h-3" viewBox="0 0 10 11" fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -98,12 +65,41 @@
 
                   <th scope="col"
                     class="px-12 py-3.5 text-sm font-normal text-left rtl:text-right text-gray-500 dark:text-gray-400">
-                    Razão Social
+                    <button class="flex items-center gap-x-3 focus:outline-none"
+                      @click="OrderingSuplier('company_name')">
+                      <span>Razão Social</span>
+
+                      <svg class="h-3" viewBox="0 0 10 11" fill="none" xmlns="http://www.w3.org/2000/svg">
+                        <path
+                          d="M2.13347 0.0999756H2.98516L5.01902 4.79058H3.86226L3.45549 3.79907H1.63772L1.24366 4.79058H0.0996094L2.13347 0.0999756ZM2.54025 1.46012L1.96822 2.92196H3.11227L2.54025 1.46012Z"
+                          fill="currentColor" stroke="currentColor" stroke-width="0.1" />
+                        <path
+                          d="M0.722656 9.60832L3.09974 6.78633H0.811638V5.87109H4.35819V6.78633L2.01925 9.60832H4.43446V10.5617H0.722656V9.60832Z"
+                          fill="currentColor" stroke="currentColor" stroke-width="0.1" />
+                        <path
+                          d="M8.45558 7.25664V7.40664H8.60558H9.66065C9.72481 7.40664 9.74667 7.42274 9.75141 7.42691C9.75148 7.42808 9.75146 7.42993 9.75116 7.43262C9.75001 7.44265 9.74458 7.46304 9.72525 7.49314C9.72522 7.4932 9.72518 7.49326 9.72514 7.49332L7.86959 10.3529L7.86924 10.3534C7.83227 10.4109 7.79863 10.418 7.78568 10.418C7.77272 10.418 7.73908 10.4109 7.70211 10.3534L7.70177 10.3529L5.84621 7.49332C5.84617 7.49325 5.84612 7.49318 5.84608 7.49311C5.82677 7.46302 5.82135 7.44264 5.8202 7.43262C5.81989 7.42993 5.81987 7.42808 5.81994 7.42691C5.82469 7.42274 5.84655 7.40664 5.91071 7.40664H6.96578H7.11578V7.25664V0.633865C7.11578 0.42434 7.29014 0.249976 7.49967 0.249976H8.07169C8.28121 0.249976 8.45558 0.42434 8.45558 0.633865V7.25664Z"
+                          fill="currentColor" stroke="currentColor" stroke-width="0.3" />
+                      </svg>
+                    </button>
                   </th>
 
                   <th scope="col"
                     class="px-4 py-3.5 text-sm font-normal text-left rtl:text-right text-gray-500 dark:text-gray-400">
-                    CNPJ
+                    <button class="flex items-center gap-x-3 focus:outline-none" @click="OrderingSuplier('document')">
+                      <span>CNPJ</span>
+
+                      <svg class="h-3" viewBox="0 0 10 11" fill="none" xmlns="http://www.w3.org/2000/svg">
+                        <path
+                          d="M2.13347 0.0999756H2.98516L5.01902 4.79058H3.86226L3.45549 3.79907H1.63772L1.24366 4.79058H0.0996094L2.13347 0.0999756ZM2.54025 1.46012L1.96822 2.92196H3.11227L2.54025 1.46012Z"
+                          fill="currentColor" stroke="currentColor" stroke-width="0.1" />
+                        <path
+                          d="M0.722656 9.60832L3.09974 6.78633H0.811638V5.87109H4.35819V6.78633L2.01925 9.60832H4.43446V10.5617H0.722656V9.60832Z"
+                          fill="currentColor" stroke="currentColor" stroke-width="0.1" />
+                        <path
+                          d="M8.45558 7.25664V7.40664H8.60558H9.66065C9.72481 7.40664 9.74667 7.42274 9.75141 7.42691C9.75148 7.42808 9.75146 7.42993 9.75116 7.43262C9.75001 7.44265 9.74458 7.46304 9.72525 7.49314C9.72522 7.4932 9.72518 7.49326 9.72514 7.49332L7.86959 10.3529L7.86924 10.3534C7.83227 10.4109 7.79863 10.418 7.78568 10.418C7.77272 10.418 7.73908 10.4109 7.70211 10.3534L7.70177 10.3529L5.84621 7.49332C5.84617 7.49325 5.84612 7.49318 5.84608 7.49311C5.82677 7.46302 5.82135 7.44264 5.8202 7.43262C5.81989 7.42993 5.81987 7.42808 5.81994 7.42691C5.82469 7.42274 5.84655 7.40664 5.91071 7.40664H6.96578H7.11578V7.25664V0.633865C7.11578 0.42434 7.29014 0.249976 7.49967 0.249976H8.07169C8.28121 0.249976 8.45558 0.42434 8.45558 0.633865V7.25664Z"
+                          fill="currentColor" stroke="currentColor" stroke-width="0.3" />
+                      </svg>
+                    </button>
                   </th>
                   <th scope="col"
                     class="px-4 py-3.5 text-sm font-normal text-left rtl:text-right text-gray-500 dark:text-gray-400">
@@ -145,16 +141,44 @@
 
                   <td class="px-4 py-4 text-sm whitespace-nowrap">
                     <div class="flex items-center sm:gap-x-5">
-                      <a href="#" @click="openModal(suplier)">
-                        <h2 class="font-medium text-gray-800 dark:text-white ">Detalhes</h2>
+                      <a href="#" @click="openModal(suplier)" class="tooltip">
+                        <h2 class="font-medium text-gray-800 dark:text-white ">
+                          <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor"
+                            class="bi bi-eye" viewBox="0 0 16 16">
+                            <path
+                              d="M16 8s-3-5.5-8-5.5S0 8 0 8s3 5.5 8 5.5S16 8 16 8zM1.173 8a13.133 13.133 0 0 1 1.66-2.043C4.12 4.668 5.88 3.5 8 3.5c2.12 0 3.879 1.168 5.168 2.457A13.133 13.133 0 0 1 14.828 8c-.058.087-.122.183-.195.288-.335.48-.83 1.12-1.465 1.755C11.879 11.332 10.119 12.5 8 12.5c-2.12 0-3.879-1.168-5.168-2.457A13.134 13.134 0 0 1 1.172 8z" />
+                            <path
+                              d="M8 5.5a2.5 2.5 0 1 0 0 5 2.5 2.5 0 0 0 0-5zM4.5 8a3.5 3.5 0 1 1 7 0 3.5 3.5 0 0 1-7 0z" />
+                          </svg>
+                        </h2>
+                        <span class="tooltiptext">Visualizar Fornecedor</span>
                       </a>
-                      <button @click="editSuplier(suplier)">
-                        <h2 class="font-medium text-gray-800 dark:text-white ">Editar</h2>
-                      </button>
-                      <a href="#" @click="deleteSuplier(suplier.id)">
-                        <h2 class="font-medium text-gray-800 dark:text-white ">Excluir</h2>
+                      <a href="#" @click="editSuplier(suplier)" class="tooltip">
+                        <h2 class="font-medium text-gray-800 dark:text-white ">
+                          <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor"
+                            class="bi bi-pencil-square" viewBox="0 0 16 16">
+                            <path
+                              d="M15.502 1.94a.5.5 0 0 1 0 .706L14.459 3.69l-2-2L13.502.646a.5.5 0 0 1 .707 0l1.293 1.293zm-1.75 2.456-2-2L4.939 9.21a.5.5 0 0 0-.121.196l-.805 2.414a.25.25 0 0 0 .316.316l2.414-.805a.5.5 0 0 0 .196-.12l6.813-6.814z" />
+                            <path fill-rule="evenodd"
+                              d="M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5v11z" />
+                          </svg>
+                        </h2>
+                        <span class="tooltiptext">Editar Fornecedor</span>
+                      </a>
+                      <a href="#" @click="deleteSuplier(suplier.id)" class="tooltip">
+                        <h2 class="font-medium text-gray-800 dark:text-white">
+                          <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor"
+                            class="bi bi-trash" viewBox="0 0 16 16">
+                            <path
+                              d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0V6Z" />
+                            <path
+                              d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1v1ZM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4H4.118ZM2.5 3h11V2h-11v1Z" />
+                          </svg>
+                        </h2>
+                        <span class="tooltiptext">Excluir Fornecedor</span>
                       </a>
                     </div>
+
                   </td>
                 </tr>
               </tbody>
@@ -266,10 +290,13 @@
 
 <script>
 import axios from 'axios';
-import FormSuplier from "@/components/TheFormSuplier.vue";
+import { ref } from 'vue';
 import { useSuplierStore } from '@/stores/suplierStore';
+import { useAlertStore } from '@/stores/alertStore';
+import FormSuplier from "@/components/TheFormSuplier.vue";
 import TheModal from '@/components/TheModal.vue';
 import Pagination from '@/components/ThePagination.vue';
+import TheAlert from '@/components/TheAlert.vue';
 
 export default {
   name: 'SuplierView',
@@ -277,6 +304,7 @@ export default {
     FormSuplier,
     TheModal,
     Pagination,
+    TheAlert,
   },
   data() {
     return {
@@ -287,6 +315,9 @@ export default {
       previous: null,
       totalPages: 10,
       currentPage: 1,
+      search: null,
+      ordering: null,
+      orderDirection: 'asc',
       contact: '',
       name: '',
       company_name: '',
@@ -298,8 +329,7 @@ export default {
       neighborhood: '',
       city: '',
       state: '',
-      successMessage: '',
-      errorMessage: '',
+      isAlertVisible: false,
       showModal: false,
       editingSuplierId: null,
       editingSuplier: {},
@@ -308,14 +338,15 @@ export default {
   },
   setup() {
     const suplierStore = useSuplierStore();
-    return { suplierStore };
+    const alertStore = useAlertStore();
+    let isAlertVisible = ref(false);
+    return { suplierStore, alertStore, isAlertVisible };
   },
   mounted() {
     this.getSuplier(this.baseURL);
   },
 
   methods: {
-
     async getSuplier(url) {
       await axios.get(url)
         .then(response => {
@@ -327,7 +358,7 @@ export default {
         .catch(error => {
           console.error('Error fetching supliers', error);
         });
-    },  
+    },
 
     loadPage(page) {
       this.currentPage = page;
@@ -336,15 +367,19 @@ export default {
     },
 
     async deleteSuplier(id) {
-      await axios.delete(`${this.baseURL}${id}/`)
-        .then(response => {
-          this.successMessage = 'Fornecedor deletado com sucesso';
-          this.getSuplier();
-        })
-        .catch(error => {
-          this.errorMessage = 'Erro ao deletar fornecedor';
+      const { alertStore } = this;
+      const isConfirmed = window.confirm('Tem certeza que deseja excluir este fornecedor?');
+      if (isConfirmed) {
+        try {
+          this.isAlertVisible = true;
+          const response = await axios.delete(`${this.baseURL}${id}/`);
+          alertStore.showAlert('Fornecedor excluido com sucesso', 'success')
+          this.getSuplier(this.baseURL);
+        } catch (error) {
+          alertStore.showAlert('Erro ao deletar fornecedor', 'error')
           console.error('Error deleting suplier', error);
-        });
+        }
+      }
     },
 
     editSuplier(suplier) {
@@ -364,6 +399,45 @@ export default {
     closeModal() {
       this.selectedSuplier = {};
       this.showModal = false;
+    },
+
+    async filterSuplier(event) {
+      if (event && event.target) {
+        this.search = event.target.value;
+        if (this.search && this.search.length >= 3) {
+          try {
+            const response = await axios.get(`${this.baseURL}?search=${this.search}`);
+            this.supliers = response.data.results;
+          } catch (error) {
+            console.error('Error filter category', error);
+          }
+        } else if (this.search === '') {
+          try {
+            const response = await axios.get(this.baseURL);
+            this.supliers = response.data.results;
+          } catch (error) {
+            console.error('Error getting all supliers', error);
+          }
+        }
+      }
+    },
+
+    async OrderingSuplier(orderField) {
+      this.orderDirection = this.orderDirection === 'asc' ? 'desc' : 'asc';
+      this.ordering = this.orderDirection === 'desc' ? `-${orderField}` : orderField;
+      await axios.get(`${this.baseURL}?ordering=${this.ordering}`)
+        .then(response => {
+          this.supliers = response.data.results;
+          if (this.orderDirection === 'asc') {
+            this.supliers.sort((a, b) => a[orderField].localeCompare(b[orderField], undefined, { sensitivity: 'accent' }));
+          } else {
+            this.supliers.sort((a, b) => b[orderField].localeCompare(a[orderField], undefined, { sensitivity: 'accent' }));
+          }
+          console.log(response.data.results);
+        })
+        .catch(error => {
+          console.error('Error ordering category', error);
+        });
     },
 
     formatContact(contact) {
